requirements.txt
customtkinter==5.2.2
Pillow==9.5.0